added storage mode parameter
