+++ commons/commons-bio/src/main/kotlin/ebi/ac/uk/model/AccNumber.kt
@@ -3,7 +3,7 @@ package ebi.ac.uk.model
 import ebi.ac.uk.base.EMPTY
 
 /**
- * Represents an submission accession number which include a pattern an a numeric value.
+ * Represents a submission accession number which include a pattern a numeric value.
  */
 data class AccNumber(val prefix: String, val numericValue: String? = null) {
     override fun toString() = "$prefix${numericValue ?: EMPTY}"
