+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/SubmitterConfig.kt
@@ -18,6 +18,7 @@ import ac.uk.ebi.biostd.submission.service.FileSourcesService
 import ac.uk.ebi.biostd.submission.service.ParentInfoService
 import ac.uk.ebi.biostd.submission.service.TimesService
 import ac.uk.ebi.biostd.submission.submitter.ExtSubmissionSubmitter
+import ac.uk.ebi.biostd.submission.submitter.SubmissionProcessor
 import ac.uk.ebi.biostd.submission.submitter.SubmissionSubmitter
 import ac.uk.ebi.biostd.submission.submitter.request.SubmissionReleaser
 import ac.uk.ebi.biostd.submission.submitter.request.SubmissionRequestLoader
@@ -89,25 +90,33 @@ class SubmitterConfig {
     @Bean
     fun submissionSubmitter(
         extSubmissionSubmitter: ExtSubmissionSubmitter,
+        submissionProcessor: SubmissionProcessor,
+        parentInfoService: ParentInfoService,
+    ) = SubmissionSubmitter(
+        extSubmissionSubmitter,
+        submissionProcessor,
+        parentInfoService,
+    )
+
+    @Bean
+    fun submissionProcessor(
         persistenceService: SubmissionPersistenceService,
         timesService: TimesService,
         accNoService: AccNoService,
         parentInfoService: ParentInfoService,
         collectionInfoService: CollectionInfoService,
-        queryService: SubmissionMetaQueryService,
         properties: ApplicationProperties,
         toExtSectionMapper: ToExtSectionMapper,
-    ) = SubmissionSubmitter(
-        extSubmissionSubmitter,
-        persistenceService,
-        timesService,
-        accNoService,
-        parentInfoService,
-        collectionInfoService,
-        queryService,
-        properties,
-        toExtSectionMapper
-    )
+    ): SubmissionProcessor =
+        SubmissionProcessor(
+            persistenceService,
+            timesService,
+            accNoService,
+            parentInfoService,
+            collectionInfoService,
+            properties,
+            toExtSectionMapper
+        )
 
     @Configuration
     class ToExtendedConfiguration {
