+++ submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/model/SubmitRequest.kt
@@ -1,5 +1,7 @@
 package ac.uk.ebi.biostd.submission.model
 
+import ebi.ac.uk.extended.model.ExtSubmission
+import ebi.ac.uk.extended.model.StorageMode
 import ebi.ac.uk.io.sources.FileSourcesList
 import ebi.ac.uk.model.Submission
 import ebi.ac.uk.model.SubmissionMethod
@@ -10,10 +12,11 @@ data class SubmitRequest(
     val submitter: SecurityUser,
     val sources: FileSourcesList,
     val method: SubmissionMethod,
-    val onBehalfUser: SecurityUser? = null,
-    val draftKey: String? = null,
+    val onBehalfUser: SecurityUser?,
+    val draftKey: String?,
+    val previousVersion: ExtSubmission?,
+    val storageMode: StorageMode?,
 ) {
     val accNo: String = submission.accNo.ifBlank { "PENDING_ACC_NO" }
-
     val owner: String = onBehalfUser?.email ?: submitter.email
 }
